src: fix the warnings the other compilers found
  * src/svm.cpp (svm_train_one): initialize si, which the compiler cannot
    see is always filled.  (solve_epsilon_svr): remove the sum the commented
    out diagnostic left behind.
  * src/libsvmwrite.cc (write): print the column index as unsigned long,
    which is what the format expects where size_t is wider.
  * src/fcnntrain.cc (fcnntrain): initialize lbres.
  * src/svm_model_octave.cc (octave_matrix_to_model): remove the id counter.

+++ src/libsvmwrite.cc
@@ -70,7 +70,7 @@ void write(string filename, ColumnVector label_vec, SparseMatrix instance_mat)
 
       for(k=low;k<high;k++)
       {
-        fprintf(fp ," %lu:%g", (size_t)ir[k]+1, samples[k]);
+        fprintf(fp ," %lu:%g", (unsigned long)ir[k]+1, samples[k]);
       }
 		  fprintf(fp, "\n");
     }
